visual_inspector_dump.py update to check for false negative flags.
Add support for new --category and --objlabel command line
parameters that take a category and object label name as
indicators of an inspection being a FALSE_NEGATIVE. If either
is present, set the FALSE_NEGATIVE field to 'True'.

Additionally, if the FALSE_NEGATIVE field is 'True', ignore
the --newerthan filter and include the inspection in the list
of inspections to be dumped.